+++ modules/update_manager/update_manager.py
@@ -45,7 +45,6 @@ class UpdateManager(Module, multiprocessing.Process):
         self.online_whitelist_update_timer = InfiniteTimer(
             self.online_whitelist_update_period, self.update_ti_files
         )
-        self.timeout = 0.000001
 
     def read_configuration(self):
         conf = ConfigParser()
@@ -116,7 +115,7 @@ class UpdateManager(Module, multiprocessing.Process):
         # Main loop function
         while True:
             try:
-                message = self.c1.get_message(timeout=self.timeout)
+                message = __database__.get_message(self.c1)
                 # Check that the message is for you. Probably unnecessary...
                 if message and message['data'] == 'stop_process':
                     self.shutdown_gracefully()