+++ slips_files/core/database/database.py
@@ -204,6 +204,22 @@ class Database(ProfilingFlowsDatabase, object):
         self.home_network = conf.get_home_network()
         self.width = conf.get_tw_width_as_float()
 
+
+    def change_redis_limits(self, redis_client):
+        """
+        To fix redis closing/resetting the pub/sub connection, change redis soft and hard limits
+        """
+        # maximum buffer size for pub/sub clients:  = 4294967296 Bytes = 4GBs,
+        # when msgs in queue reach this limit, Redis will
+        # close the client connection as soon as possible.
+
+        # soft limit for pub/sub clients: 2147483648 Bytes = 2GB over 10 mins,
+        # means if the client has an output buffer bigger than 2GB
+        # for, continuously, 10 mins, the connection gets closed.
+        redis_client.config_set('client-output-buffer-limit', "normal 0 0 0 "
+                                                        "slave 268435456 67108864 60 "
+                                                        "pubsub 4294967296 2147483648 600")
+
     def start(self, redis_port):
         """Start the DB. Allow it to read the conf"""
         self.read_configuration()
@@ -225,13 +241,15 @@ class Database(ProfilingFlowsDatabase, object):
                     # to close slips files
                     self.r.flushdb()
 
-                self.r.config_set('client-output-buffer-limit', "normal 0 0 0 slave 268435456 67108864 60 pubsub 1073741824 1073741824 600")
-                self.rcache.config_set('client-output-buffer-limit', "normal 0 0 0 slave 268435456 67108864 60 pubsub 1073741824 1073741824 600")
+                self.change_redis_limits(self.r)
+                self.change_redis_limits(self.rcache)
+
                 # to fix redis.exceptions.ResponseError MISCONF Redis is configured to save RDB snapshots
                 # configure redis to stop writing to dump.rdb when an error occurs without throwing errors in slips
                 # Even if the DB is not deleted. We need to delete some temp data
-                # Zeek_files
                 self.r.delete('zeekfiles')
+
+
             # By default the slips internal time is 0 until we receive something
             self.setSlipsInternalTime(0)
             while self.get_slips_start_time() is None:
@@ -239,6 +257,31 @@ class Database(ProfilingFlowsDatabase, object):
         except redis.exceptions.ConnectionError as ex:
             print(f"[DB] Can't connect to redis on port {redis_port}: {ex}")
             return False
+
+    def is_connection_error_logged(self):
+        return True if self.r.get('logged_connection_error') else False
+
+    def mark_connection_error_as_logged(self):
+        """
+        When redis connection error occurs, to prevent every module from logging it to slips.log and the console,
+        set this variable in the db
+        """
+        self.r.set('logged_connection_error', 'True')
+
+    def get_message(self, channel, timeout=0.0000001):
+        """
+        Wrapper for redis' get_message() to be able to handle redis.exceptions.ConnectionError
+        notice: there has to be a timeout or the channel will wait forever and never receive a new msg
+        """
+        try:
+            return channel.get_message(timeout=timeout)
+        except redis.exceptions.ConnectionError as ex:
+            if not self.is_connection_error_logged():
+                self.publish('finished_modules', 'stop_slips')
+                self.print(f'Stopping slips due to redis.exceptions.ConnectionError: {ex}',0,1)
+                # make sure we publish the stop msg and log the error only once
+                self.mark_connection_error_as_logged()
+
 
     def set_slips_start_time(self):
         """store the time slips started (datetime obj)"""
@@ -1662,7 +1705,10 @@ class Database(ProfilingFlowsDatabase, object):
         return self.pubsub
 
     def publish_stop(self):
-        """Publish stop command to terminate slips"""
+        """
+        Publish stop command to terminate slips
+        to shutdown slips gracefully, this function should only be used by slips.py
+        """
         all_channels_list = self.r.pubsub_channels()
         self.print('Sending the stop signal to all listeners', 0, 3)
         for channel in all_channels_list: