+++ CHANGELOG.md
@@ -17,6 +17,9 @@
 - whitelist top tranco top 10k domains for fewer false positive alerts
 - Detect ICMP scans in netflow files
 - Fix P2P
+- Kill slips on  when redis ConnectionError occurs
+- Kill all modules after 15 mins to trying to stop them
+- Keep track of profiles' past threat levels
 - Add an option to store the zeek log files inside the output dir  
 
 -0.9.5